% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{aggregate.EM}
\alias{aggregate.EM}
\title{Aggregate EM objects for a single summary}
\usage{
aggregate.EM(object)
}
\arguments{
\item{object}{A list of lists, where each sub-list contains three entries:
EM (the EM object), nobs_a (the number of observations in dataset A) and
nobs_b (the number of observations in dataset B)}
}
\description{
\code{aggregate.EM} aggregates EM objects to create a single statewide summary.
}

