% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{aggregateEM}
\alias{aggregateEM}
\title{Aggregate EM objects for a single summary}
\usage{
aggregateEM(object)
}
\arguments{
\item{object}{A list of lists, where each sub-list contains three entries:
EM (the EM object), nobs_a (the number of observations in dataset A) and
nobs_b (the number of observations in dataset B)}
}
\description{
\code{aggregateEM} aggregates EM objects to create a single statewide summary.
}

