% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockData.R
\name{blockData}
\alias{blockData}
\title{blockData}
\usage{
blockData(dfA, dfB, varnames, window.block, window.size,
kmeans.block, nclusters, iter.max,
stringdist.subset, similarity.threshold, stringdist.method, jw.weight,
n.cores)
}
\arguments{
\item{dfA}{Dataset A - to be matched to Dataset B}

\item{dfB}{Dataset B - to be matched to Dataset A}

\item{varnames}{A vector of variable names to use for blocking.
Must be present in both dfA and dfB}

\item{window.block}{A vector of variable names indicating that the variable should be
blocked using windowing blocking. Must be present in varnames.}

\item{window.size}{The size of the window for window blocking. Default is 1
(observations +/- 1 on the specified variable will be blocked together).}

\item{kmeans.block}{A vector of variable names indicating that the variable should be
blocked using k-means blocking. Must be present in varnames.}

\item{nclusters}{Number of clusters to create with k-means. Default value is the
number of clusters where the average cluster size is 100,000 observations.}

\item{iter.max}{Maximum number of iterations for the k-means algorithm to run. Default is 5000}

\item{stringdist.subset}{A single variable name present in dfA and dfB to run string-distance
subsetting on, so that any observations that is not close to any other observation on the metric
will be excluded from all blocks.}

\item{similarity.threshold}{Lower bound on string-distance measure for being considered a possible match.
If an observation has no possible matches above this threshold, it is discarded from the match. Default is 0.8.}

\item{stringdist.method}{The method to use for calculating string-distance similarity. Possible values are
'jaro' (Jaro Distance), 'jw' (Jaro-Winkler), and 'lv' (Levenshtein). Default is 'jw'.}

\item{jw.weight}{Parameter that describes the importance of the first characters of a string (only needed if stringdist.method = "jw"). Default is .10.}

\item{n.cores}{Number of cores to parallelize over. Default is NULL.}
}
\value{
A list with an entry for each block. Each list entry contains two vectors --- one with the indices indicating the block members in dataset A,
and another containing the indices indicating the block members in dataset B.
}
\description{
Contains functionalities for blocking two data sets on one or more variables prior to
conducting a merge.
}
