% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPriors.R
\name{calcPriors}
\alias{calcPriors}
\title{calcPriors}
\usage{
calcPriors(state_a, state_b, year_start, year_end, var)
}
\arguments{
\item{state_a}{The state code for the earlier of the two voter files.}

\item{state_b}{The state code for the later of the two voter files.}

\item{year_start}{The year of the voter file for state A.}

\item{year_end}{The year of the voter file for state B.}

\item{var}{User-specified variance for the prior probability.}
}
\description{
calcPriors estimates optimal \eqn{\alpha} and \eqn{\beta} values
for the Beta prior on \eqn{\gamma}, and optimal \eqn{\alpha_1} and
\eqn{\alpha_2} values for the Dirichlet prior on \eqn{\pi_{j,k}} when matching
state voter files over time.
}
\author{
Ben Fifield <benfifield@gmail.com>
}

