% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CKpar.b.R
\name{CKpar.b}
\alias{CKpar.b}
\title{CKpar.b}
\usage{
CKpar.b(matAp, matBp, blocklist, dedupe.blocks, n.cores, transform, transform.args, 
method, method.args, isoreg, prior, resolution)
}
\arguments{
\item{matAp}{vector storing the comparison field in data set 1}

\item{matBp}{vector storing the comparison field in data set 2}

\item{blocklist}{list of blocking indices}

\item{n.cores}{Number of cores to parallelize over. Default is NULL.}

\item{transform}{An optional function returning transformed strings or a list of transformed strings. 
First argument should be the strings to be transformed}

\item{transform.args}{An optional list of arguments to the transform function}

\item{method}{String distance method, passed as a custom function with argument list stringdist.args. The first two arguments of 
any custom function provided should be the strings to be compared}

\item{method.args}{List of arguments for custom function provided to method}

\item{dedupe.blocks}{Logical indicators for whether internal linkage and duplication is taking place}

\item{isoreg}{Isotonic regression object that will be scaled by prior weights to predict conditional probability of matching based on similarity scores}

\item{priors}{prior probability of match within the stratum examined for each block}

\item{resolution}{Number of probability bins over which to aggregate results}

\item{drop.lowest}{Whether to drop matched indices from the lowest probability bin - this can save a lot of memory and processing wasted on nonmatches}
}
\value{
\code{CKpar.b} returns a list of lists (for each block) with the indices corresponding to posterior match probabilities within a number of probability bins given by the resolution parameter
}
\description{
posterior match probability calculation based on string similarity and Fellegi-Sunter priors
}
