% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanAddressUSPS.R
\name{cleanAddressUSPS}
\alias{cleanAddressUSPS}
\title{cleanAddressUSPS}
\usage{
cleanAddressUSPS(address.field)
}
\arguments{
\item{address.field}{A vector containing address information to be cleaned.}
}
\value{
\code{cleanAddressUSPS()} returns a cleaned version of \code{address.field}.
}
\description{
Apply USPS address standardization to address field.
}
\examples{
dfA$streetname <- cleanAddressUSPS(dfA$streetname)
}
\author{
Ted Enamorado <ted.enamorado@gmail.com> and Ben Fifield <benfifield@gmail.com>
}

