% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterWordEmbed.R
\name{clusterWordEmbed}
\alias{clusterWordEmbed}
\title{clusterWordEmbed}
\usage{
clusterWordEmbed(vecA, vecB, nclusters, max_n, k)
}
\arguments{
\item{vecA}{The character vector from dataset A}

\item{vecB}{The character vector from dataset B}

\item{nclusters}{The number of clusters to create from the provided data. Either
nclusters = NULL or max_n = NULL.}

\item{max_n}{The maximum size of either dataset A or dataset B in
the largest cluster. Either nclusters = NULL or max_n = NULL}

\item{k}{The number of dimensions to use from PCA for k-means.}
}
\value{
\code{clusterWordEmbed} returns a list of length 3:
\item{clusterA}{The cluster assignments for dataset A}
\item{clusterB}{The cluster assignments for dataset B}
\item{n.clusters}{The number of clusters created}
}
\description{
Use word embedding to create properly sized clusters for matching.
The function first creates a word embedding out of the provided
vectors, and then runs PCA on the matrix. It then takes the first
\code{k} dimensions (where \code{k} is provided by the user) and
k-means is run on that matrix to get the clusters.
}
\examples{
data(samplematch)
cl <- clusterWordEmbed(dfA$firstname, dfB$firstname, nclusters = 3)
}
\author{
Ben Fifield <benfifield@gmail.com>
}

