% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockData.R
\name{consolidate_blocks}
\alias{consolidate_blocks}
\title{consolidate_blocks}
\usage{
consolidate_blocks(blocklist1, blocklist2, ..., dedupe = F)
}
\arguments{
\item{...}{Any number of blocklists output by \code{blockData()}, separated by commas}

\item{dedupe}{Boolean indicator as to whether the blocks represent an internal linkage problem (\code{TRUE}) or linkage between two datasets (\code{FALSE})}
}
\value{
A list of non-overlapping blocks covering the same record pairs as the original inputs
}
\description{
Allows for disjunctive blocking using multiple criteria, i.e. ONE OR MORE blocking conditions must be met
}
