% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countyfips}
\alias{countyfips}
\title{County-level FIPS Codes}
\format{A dataframe containing 3235 observations.}
\usage{
countyfips
}
\description{
This data maps county names to FIPS codes for
use in calculating prior movers rates.
}
\keyword{dataset}

