% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countyinflow}
\alias{countyinflow}
\title{County-level inflow rates by state}
\format{A dataframe containing 423752 observations.}
\usage{
countyinflow
}
\description{
This data compiles and cleans county-level
movers inflow rates by county, from the IRS
Statistics on Income dataset.
}
\keyword{dataset}
