% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countyoutflow}
\alias{countyoutflow}
\title{County-level outflow rates by state}
\format{A dataframe containing 424475 observations.}
\usage{
countyoutflow
}
\description{
This data compiles and cleans county-level
movers outflow rates by county, from the IRS
Statistics on Income dataset.
}
\keyword{dataset}
