% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedupeMatches.R
\name{dedupeMatches}
\alias{dedupeMatches}
\title{dedupeMatches}
\usage{
dedupeMatches(matchesA, matchesB, EM,
matchesLink, varnames, stringdist.match, partial.match,
cut.a = 0.92, cut.p = 0.88)
}
\arguments{
\item{matchesA}{A dataframe of the matched observations in
dataset A, with all variables used to inform the match.}

\item{matchesB}{A dataframe of the matched observations in
dataset B, with all variables used to inform the match.}

\item{EM}{The EM object from \code{emlinkMARmov()}}

\item{matchesLink}{The output from \code{matchesLink()}}

\item{varnames}{A vector of variable names to use for matching.
Must be present in both matchesA and matchesB.}

\item{stringdist.match}{A vector of booleans, indicating whether to use
string distance matching when determining matching patterns on
each variable. Must be same length as varnames.}

\item{partial.match}{A vector of booleans, indicating whether to include
a partial matching category for the string distances. Must be same length
as varnames. Default is FALSE for all variables.}

\item{cut.a}{Lower bound for full string-distance match, ranging between 0 and 1. Default is 0.92}

\item{cut.p}{Lower bound for partial string-distance match, ranging between 0 and 1. Default is 0.88}
}
\value{
\code{dedupeMatches()} returns a list containing the following elements:
\item{matchesA}{A deduped version of matchesA}
\item{matchesB}{A deduped version of matchesB}
\item{EM}{A deduped version of the EM object}
}
\description{
Dedupe matched dataframes using the linear sum assignment problem
methodology.
}
\author{
Ted Enamorado <ted.enamorado@gmail.com> and Ben Fifield <benfifield@gmail.com>
}

