% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deletion.R
\name{deletion}
\alias{deletion}
\title{deletion}
\usage{
deletion(x)
}
\arguments{
\item{x}{object that holds the string variable (value) to be modified.}
}
\description{
This is a function that deletes a character at random from a string.
It was created as a form to add noise to a field that
holds a string variable in the validation exercises that can be found
in Enamorado and Imai (Nd)
}
\author{
Ted Enamorado <ted.enamorado@gmail.com> and Kosuke Imai
}
