% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dfA}
\alias{dfA}
\title{Sample dataset A}
\format{
A dataframe containing 500 observations.
}
\usage{
dfA
}
\description{
This data is a randomized and anonymized sample dataset
to display features of fastLink.
}
\keyword{dataset}
