% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dfB}
\alias{dfB}
\title{Sample dataset B}
\format{A dataframe containing 350 observations.}
\usage{
dfB
}
\description{
This data is a randomized and anonymized sample dataset
to display features of fastLink.
}
\keyword{dataset}

