% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emlinkMAR.R
\name{emlinkMAR}
\alias{emlinkMAR}
\title{emlinkMAR}
\usage{
emlinkMAR(patterns, p.m = NULL, p.gamma.k.m = NULL, p.gamma.k.u = NULL,
  tol = NULL, iter.max = NULL)
}
\arguments{
\item{patterns}{table that holds the counts for each unique agreement
pattern. This object is produced by the function: tableCounts.}

\item{p.m}{probability of finding a match}

\item{p.gamma.k.m}{probability that conditional of being in the matched set we observed a specific agreement value for field k.}

\item{p.gamma.k.u}{probability that conditional of being in the non-matched set we observed a specific agreement value for field k.}

\item{tol}{convergence tolerance}

\item{iter.max}{Max Number of Iterations (5000 by default)}
}
\description{
Expectation-Maximization algorithm for Record Linkage under the
Missing at Random (MAR) assumption.
}
\author{
Ted Enamorado <ted.enamorado@gmail.com> and Kosuke Imai
}

