% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emlinkMARmov.R
\name{emlinkRS}
\alias{emlinkRS}
\title{emlinkRS}
\usage{
emlinkRS(patterns.out, em.out, nobs.a, nobs.b)
}
\arguments{
\item{patterns.out}{The output from `tableCounts()` or `emlinkMARmov()` (run on full dataset),
containing all observed matching patterns in the full sample and the number of times that pattern
is observed.}

\item{em.out}{The output from `emlinkMARmov()`, an EM object estimated
on a smaller random sample to apply to counts from a larger sample}

\item{nobs.a}{Total number of observations in dataset A}

\item{nobs.b}{Total number of observations in dataset B}
}
\description{
Calculates Felligi-Sunter weights and posterior zeta probabilities
for matching patterns observed in a larger population that are
not present in a sub-sample used to estimate the EM.
}
\author{
Ted Enamorado <ted.enamorado@gmail.com> and Ben Fifield <benfifield@gmail.com>
}

