% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastLink-package.R
\docType{package}
\name{fastLink-package}
\alias{fastLink}
\alias{fastLink-package}
\title{Record Linkage under MAR}
\description{
This packages provides functions to perform link (merge
probabilistically) two datasets when a unique identifier is missing.
}
\details{
\tabular{ll}{ Package: \tab fastLink\cr Type: \tab Package\cr Version: \tab 0.1.-\cr
Date: \tab 2017-01-27\cr License: \tab GPL (>= 3)\cr }
}
\author{
Ted Enamorado \email{tede@princeton.edu} and Kosuke Imai \email{kimai@princeton.edu}

Maintainer: Ted Enamorado \email{tede@princeton.edu}
}
\keyword{package}

