% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_dependencies.R
\name{find_dependencies}
\alias{find_dependencies}
\title{find_dependencies}
\usage{
find_dependencies(f)
}
\arguments{
\item{f}{the function to be analyzed}

\item{drops}{function calls to ignore. Used to limit redundancy during
recursion}

\item{lyr}{the current level of recursion, used to control warning messages}
}
\value{
\code{find_dependencies} returns a list with members 'depends', a
  data table containing packages and functions called by the target, and
  'missing', a vector of functions referred to within the call stack that are
  not currently defined. Note that missing functions are present in the call
  stack, but may not actually be called depending on other function
  parameters
}
\description{
get functions and packages required by a function. Used for parallelization
of custom string comparison functions
}
