% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammaCK2par.b.R
\name{gammaCK2par.b}
\alias{gammaCK2par.b}
\title{gammaCK2par.b}
\usage{
gammaCK2par.b(matAp, matBp, blocklist, dedupe.blocks, n.cores, cut.a, transform, transform.args, method, method.args, w)
}
\arguments{
\item{matAp}{vector storing the comparison field in data set 1}

\item{matBp}{vector storing the comparison field in data set 2}

\item{blocklist}{list of blocking indices}

\item{n.cores}{Number of cores to parallelize over. Default is NULL.}

\item{cut.a}{Lower bound for full match. Default is 0.92}

\item{transform}{An optional function returning transformed strings or a list of transformed strings. 
First argument should be the strings to be transformed}

\item{transform.args}{An optional list of arguments to the transform function}

\item{method}{String distance method, options are: "jw" Jaro-Winkler (Default), "jaro" Jaro, and "lv" Edit. 
May also be passed as a custom function with argument list stringdist.args. The first two arguments of 
any custom function provided should be the strings to be compared}

\item{w}{Parameter that describes the importance of the first characters of a string (only needed if method = "jw"). Default is .10}

\item{method.args}{List of arguments for custom function provided to method}

\item{dedupe.blocks}{Logical indicators for whether internal linkage and duplication is taking place}
}
\value{
\code{gammaCK2par.b} returns a list of lists (for each block) with the indices corresponding to each
matching pattern, which can be fed directly into \code{tableCounts.b} and \code{matchesLink.b}.
}
\description{
Field comparisons for string variables for blocked data. Two possible agreement patterns are considered:
0 total disagreement, 2 agreement.
Distance between strings may be calculated using various metrics, including custom functions
}
\author{
Ted Enamorado <ted.enamorado@gmail.com>, Ben Fifield <benfifield@gmail.com>, and Kosuke Imai
}
