% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammaCK2par.R
\name{gammaCK2par}
\alias{gammaCK2par}
\title{gammaCK2par}
\usage{
gammaCK2par(matAp, matBp, n.cores = NULL, cut.a = NULL,
  calc.prior = FALSE, var = NULL)
}
\arguments{
\item{matAp}{vector storing the comparison field in data set 1}

\item{matBp}{vector storing the comparison field in data set 2}
}
\description{
Field comparisons for string variables. Two possible agreement patterns are considered:
0 total disagreement, 2 agreement.
The distance between strings is calculated using a Jaro-Winkler distance as implemented
in the stringdist package.
}
\author{
Ted Enamorado <ted.enamorado@gmail.com>, Ben Fifield <benfifield@gmail.com>, and Kosuke Imai
}

