% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammaKpar.b.R
\name{gammaKpar.b}
\alias{gammaKpar.b}
\title{gammaKpar.b}
\usage{
gammaKpar.b(matAp, matBp, blocklist, gender, dedupe.blocks, n.cores)
}
\arguments{
\item{matAp}{vector storing the comparison field in data set 1}

\item{matBp}{vector storing the comparison field in data set 2}

\item{blocklist}{list of blocking indices}

\item{gender}{Whether the matching variable is gender. Will override
standard warnings of missingness/nonvariability. Default is FALSE.}

\item{dedupe.blocks}{Logical indicators for whether internal linkage and duplication is taking place}

\item{n.cores}{Number of cores to parallelize over. Default is NULL.}
}
\value{
\code{gammaKpar.b} returns a list (over blocks) of lists with the indices corresponding to each
matching pattern, which can be fed directly into \code{tableCounts.b} and \code{matchesLink.b}.
}
\description{
Field comparisons for blocked data: 0 disagreement, 2 total agreement.
}
\author{
Ted Enamorado <ted.enamorado@gmail.com>, Ben Fifield <benfifield@gmail.com>, and Kosuke Imai
}
