% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammaKpar.R
\name{gammaKpar}
\alias{gammaKpar}
\title{gammaKpar}
\usage{
gammaKpar(matAp, matBp, n.cores = NULL)
}
\arguments{
\item{matAp}{vector storing the comparison field in data set 1}

\item{matBp}{vector storing the comparison field in data set 2}
}
\description{
Field comparisons: 0 disagreement, 2 total agreement.
}
\author{
Ted Enamorado <ted.enamorado@gmail.com>, Ben Fifield <benfifield@gmail.com>, and Kosuke Imai
}

