% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammaNUMCK2par.b.R
\name{gammaNUMCK2par.b}
\alias{gammaNUMCK2par.b}
\title{gammaNUMCK2par.b}
\usage{
gammaNUMCK2par.b(matAp, matBp, blocklist, dedupe.blocks, n.cores, cut.a)
}
\arguments{
\item{matAp}{vector storing the comparison field in data set 1}

\item{matBp}{vector storing the comparison field in data set 2}

\item{blocklist}{list of blocking indices}

\item{n.cores}{Number of cores to parallelize over. Default is NULL.}

\item{cut.a}{Lower bound for full match. Default is 1}

\item{dedupe.blocks}{Logical indicators for whether internal linkage and duplication is taking place}
}
\value{
\code{gammaNUMCK2par.b} returns a list of lists (for each block) with the indices corresponding to each
matching pattern, which can be fed directly into \code{tableCounts.b} and \code{matchesLink.b}.
}
\description{
Field comparisons for numeric variables for blocked data. Three possible agreement patterns are considered:
0 total disagreement, 2 agreement.
The distance between numbers is calculated using their absolute distance.
}
\author{
Ted Enamorado <ted.enamorado@gmail.com>, Ben Fifield <benfifield@gmail.com>, and Kosuke Imai
}
