% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammaNUMCK2par.R
\name{gammaNUMCK2par}
\alias{gammaNUMCK2par}
\title{gammaNUMCK2par}
\usage{
gammaNUMCK2par(matAp, matBp, n.cores, cut.a)
}
\arguments{
\item{matAp}{vector storing the comparison field in data set 1}

\item{matBp}{vector storing the comparison field in data set 2}

\item{n.cores}{Number of cores to parallelize over. Default is NULL.}

\item{cut.a}{Lower bound for full match. Default is 1}
}
\value{
\code{gammaNUMCK2par} returns a list with the indices corresponding to each
matching pattern, which can be fed directly into \code{tableCounts} and \code{matchesLink}.
}
\description{
Field comparisons for numeric variables. Two possible agreement patterns are considered:
0 total disagreement, 2 agreement.
The distance between numbers is calculated using their absolute distance.
}
\examples{
\dontrun{
g1 <- gammaNUMCK2par(dfA$birthyear, dfB$birthyear)
}
}
\author{
Ted Enamorado <ted.enamorado@gmail.com>, Ben Fifield <benfifield@gmail.com>, and Kosuke Imai
}
