% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammaNUMCKpar.R
\name{gammaNUMCKpar}
\alias{gammaNUMCKpar}
\title{gammaNUMCKpar}
\usage{
gammaNUMCKpar(matAp, matBp, n.cores, cut.a, cut.p)
}
\arguments{
\item{matAp}{vector storing the comparison field in data set 1}

\item{matBp}{vector storing the comparison field in data set 2}

\item{n.cores}{Number of cores to parallelize over. Default is NULL.}

\item{cut.a}{Lower bound for full match. Default is 1}

\item{cut.p}{Lower bound for partial match. Default is 2}
}
\value{
\code{gammaNUMCKpar} returns a list with the indices corresponding to each
matching pattern, which can be fed directly into \code{tableCounts} and \code{matchesLink}.
}
\description{
Field comparisons for numeric variables. Three possible agreement patterns are considered:
0 total disagreement, 1 partial agreement, 2 agreement.
The distance between numbers is calculated using their absolute distance.
}
\examples{
\dontrun{
g1 <- gammaNUMCKpar(dfA$birthyear, dfB$birthyear)
}
}
\author{
Ted Enamorado <ted.enamorado@gmail.com>, Ben Fifield <benfifield@gmail.com>, and Kosuke Imai
}
