% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectEM.R
\name{inspectEM}
\alias{inspectEM}
\title{inspectEM}
\usage{
inspectEM(object, posterior.range, digits)
}
\arguments{
\item{object}{The output from either \code{fastLink} or \code{emlinkMARmov}.}

\item{posterior.range}{The range of posterior probabilities to display.
Default is c(0.85, 1).}

\item{digits}{How many digits to include in inspectEM dataframe. Default is 3.}
}
\value{
\code{inspectEM} returns a data frame with information
about patterns around the provided threshold.
}
\description{
Inspect EM objects to analyze successfully and
unsuccessfully matched patterns.
}
\author{
Ben Fifield <bfifield@princeton.edu>
}

