% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchesLink.b.R
\name{matchesLink.b}
\alias{matchesLink.b}
\title{matchesLink.b}
\usage{
matchesLink.b(gammalist, blocklist, em, thresh, dedupe.blocks, n.cores = NULL)
}
\arguments{
\item{gammalist}{A list of objects produced by either gammaKpar.b or
gammaCKpar.b.}

\item{blocklist}{A list containing indices of blocks for record linkage.}

\item{em}{parameters obtained from the Expectation-Maximization algorithm under the MAR assumption. These estimates are
produced by emlinkMARmov}

\item{thresh}{is the interval of posterior zeta values for the agreements that we want to examine closer. Ranges between 0 and 1.
Can be a vector of length 1 (from specified value to 1) or 2 (from first specified value to second specified value).}

\item{dedupe.blocks}{vector of logical indicators for whether a block represents a self-comparison set of data}

\item{n.cores}{Number of cores to parallelize over. Default is NULL.}
}
\value{
\code{matchesLink.b} returns an nmatches X 2 matrix with the indices of the
matches rows in dataset A and dataset B.
}
\description{
matchesLink.b, operating on blocked data, produces two dataframes that store
all the pairs that share a pattern that conforms
to the an interval of the Fellegi-Sunter
weights
}
\author{
Ted Enamorado <ted.enamorado@gmail.com>, Ben Fifield <benfifield@gmail.com>, and Kosuke Imai
}
