% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.fastLink}
\alias{plot.fastLink}
\title{Plot matching patterns of the EM object by posterior probability of match}
\usage{
\method{plot}{fastLink}(x, posterior.range, ...)
}
\arguments{
\item{x}{Either a \code{fastLink} or \code{fastLink.EM} object to be plotted.}

\item{posterior.range}{The range of posterior probabilities to display.
Default is c(0.85, 1).}

\item{...}{Further arguments to be passed to \code{plot.fastLink()}.}
}
\description{
\code{plot.fastLink()} plots the matching patterns of the EM object,
ordering the matching patterns by the posterior probability of the match.
}
