% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMoversPriors.R
\name{precalcPriors}
\alias{precalcPriors}
\title{precalcPriors}
\usage{
precalcPriors(L, var.prior.lambda = NULL, var.prior.pi = NULL,
  lambda.mean = NULL, pi.mean = NULL, max.iter = 100)
}
\arguments{
\item{L}{Number of agreement categories for \eqn{\pi_{k,l}}. Default is NULL.}

\item{var.prior.lambda}{User-specified variance for the prior probability on lambda.}

\item{var.prior.pi}{User-specified variance for the prior probability on pi.}

\item{lambda.mean}{Prior mean for \eqn{\lambda}}

\item{pi.mean}{Prior mean for \eqn{pi_{k,l}}}

\item{max.iter}{Maximum powers of 10 that should be tried to find a proper
variance for priors before failing. Default is 100.}
}
\author{
Ben Fifield <benfifield@gmail.com>
}
