% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocText.R
\name{preprocText}
\alias{preprocText}
\title{preprocText}
\usage{
preprocText(text, convert_text, tolower, soundex,
usps_address, convert_text_to)
}
\arguments{
\item{text}{A vector of text data to convert.}

\item{convert_text}{Whether to convert text to the desired encoding, where
the encoding is specified in the 'convert_text_to' argument. Default is
TRUE}

\item{tolower}{Whether to normalize the text to be all lowercase. Default is
TRUE.}

\item{soundex}{Whether to convert the field to the Census's soundex encoding.
Default is FALSE.}

\item{usps_address}{Whether to use USPS address standardization rules to clean address fields.
Default is FALSE.}

\item{convert_text_to}{Which encoding to use when converting text. Default is 'Latin-ASCII'.
Full list of encodings in the \code{stri_trans_list()} function in the \code{stringi} package.}
}
\value{
\code{preprocText()} returns the preprocessed vector of text.
}
\description{
Preprocess text data such as names and addresses.
}
\author{
Ben Fifield <benfifield@gmail.com>
}
