% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.inspectEM}
\alias{print.inspectEM}
\title{print.inspectEM}
\usage{
\method{print}{inspectEM}(x, ...)
}
\arguments{
\item{x}{An \code{inspectEM} object}

\item{...}{Further arguments to be passed to \code{print.fastLink()}.}
}
\description{
Print information from the EM algorithm to console.
}

