% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{statefips}
\alias{statefips}
\title{State-level FIPS Codes}
\format{A dataframe containing 54 observations.}
\usage{
statefips
}
\description{
This data maps state names to FIPS codes for
use in calculating prior movers rates.
}
\keyword{dataset}

