% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stateinflow}
\alias{stateinflow}
\title{State-level inflow rates by state}
\format{A dataframe containing 11321 observations.}
\usage{
stateinflow
}
\description{
This data compiles and cleans state-level
movers inflow rates by state, from the IRS
Statistics on Income dataset.
}
\keyword{dataset}

