% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{statemove}
\alias{statemove}
\title{In-state movers rates by state}
\format{
A dataframe containing 51 observations.
}
\usage{
statemove
}
\description{
This data collects in-state movers rates by state,
for imputation where within-county movers rates are
not available.
}
\keyword{dataset}
