% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stateoutflow}
\alias{stateoutflow}
\title{State-level outflow rates by state}
\format{
A dataframe containing 11320 observations.
}
\usage{
stateoutflow
}
\description{
This data compiles and cleans state-level
movers outflow rates by state, from the IRS
Statistics on Income dataset.
}
\keyword{dataset}
