% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastLinkSB_wrap.R
\name{svblinkSB}
\alias{svblinkSB}
\title{svblinkSB}
\usage{
svblinkSB(dfA, dfB, nA, nB, n_block, tol, Lk, iter.max)
}
\description{
Stochastic Variational Inference for Record Linkage with Stochastic Blocking.
}
\details{
This function implements stochastic variational inference (with mini-batch) for parameter estimation 
 for Stochastic Blocking Fellegi-Sunter Model (SBFS).
}
\seealso{
\code{\link{fastLink}} for fast implementation of Fellegi-Sunter model,
 and \code{\link{fastLinkSB}} for the full procedure.
}
\author{
Soichiro Yamauchi \email{syamauchi@princeton.edu}
}
