% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableCounts.b.R
\name{tableCounts.b}
\alias{tableCounts.b}
\title{tableCounts.b}
\usage{
tableCounts.b(gammalist, blocklist, dedupe.blocks, n.cores)
}
\arguments{
\item{gammalist}{A list (across blocks) of lists of objects produced by gammaKpar, gammaCK2par, or
gammaCKpar.}

\item{blocklist}{A list of indices for the blocks}

\item{dedupe.blocks}{vector of logical indicators for whether a block represents self-comparison of a set of data}

\item{n.cores}{Number of cores to parallelize over. Default is NULL.}
}
\value{
\code{tableCounts.b} returns counts of all unique mathching patterns, which can be
fed directly into \code{emlinkMAR} to get posterior matching probabilities for each unique pattern.
}
\description{
Count pairs with the same pattern in the cross product between two datasets. Version for blocked data.
}
\author{
Ted Enamorado <ted.enamorado@gmail.com>, Ben Fifield <benfifield@gmail.com>, and Kosuke Imai
}
