% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpos.R
\name{transpos}
\alias{transpos}
\title{transpos}
\usage{
transpos(x)
}
\arguments{
\item{x}{object that holds the string variable (value) to be modified.}
}
\description{
This is a function that transposes (chages its order of appeareance) a
character at random from a string. It was created as a form to add noise
to a field that holds a string variable in the validation exercises that
can be found in Enamorado and Imai (Nd)
}
\author{
Ted Enamorado <ted.enamorado@gmail.com> and Kosuke Imai
}
