% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typo.R
\name{typo}
\alias{typo}
\title{typo}
\usage{
typo(x)
}
\arguments{
\item{x}{object that holds the string variable (value) to be modified.}
}
\description{
This is a function that creates a typo within a string. It was created as
a form to add noise to a field that holds a string variable in the validation
exercises that can be found in Enamorado and Imai (Nd)
}
\author{
Ted Enamorado <ted.enamorado@gmail.com> and Kosuke Imai
}

