% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggconfusion.R
\name{aggconfusion}
\alias{aggconfusion}
\title{aggconfusion}
\usage{
aggconfusion(object)
}
\arguments{
\item{object}{A list of confusion tables.}
}
\value{
'aggconfusion()' returns two tables - one calculating the confusion table, and another
calculating a series of additional summary statistics.
}
\description{
Aggregate confusion tables from separate runs of fastLink()
}
\author{
Ted Enamorado <ted.enamorado@gmail.com> and Ben Fifield <benfifield@gmail.com>
}
