% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{aggregateEM}
\alias{aggregateEM}
\title{Aggregate EM objects for use in `summary.fastLink()`}
\usage{
aggregateEM(em.list, within.geo)
}
\arguments{
\item{em.list}{A list of `fastLink` or `fastLink.EM` objects that should be aggregate
in `summary.fastLink()`}

\item{within.geo}{A vector of booleans corresponding to whether each object in `em.list`
is a within-geography match or an across-geography match. Should be of equal length to
`em.list`. Default is NULL (assumes all are within-geography matches).}
}
\description{
\code{aggregateEM} aggregates EM objects for easy processing by `summary.fastLink()`
}
