% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPriors.R
\name{calcPriors}
\alias{calcPriors}
\title{calcPriors}
\usage{
calcPriors(geo_a, geo_b, year_start, year_end, var_prior, L = NULL,
  county = FALSE, state_a = NULL, state_b = NULL, denom_mu = NULL)
}
\arguments{
\item{geo_a}{The state code (if state = TRUE) or county name
(if state = FALSE) for the earlier of the two voter files.}

\item{geo_b}{The state code (if state = TRUE) or county name
(if state = FALSE) for the later of the two voter files.}

\item{year_start}{The year of the voter file for geography A.}

\item{year_end}{The year of the voter file for geography B.}

\item{var_prior}{User-specified variance for the prior probability.}

\item{L}{Number of agreement categories for \eqn{\pi_{k,l}}. Default is NULL.}

\item{county}{Whether prior is being calculated on the county or state level.
Default is FALSE (for a state-level calculation).}

\item{state_a}{If \code{county = TRUE} (indicating a county-level match),
the state code of \code{geo_a}. Default is NULL.}

\item{state_b}{If \code{county = TRUE} (indicating a county-level match),
the state code of \code{geo_b}. Default is NULL.}
}
\description{
calcPriors estimates optimal \eqn{\alpha} and \eqn{\beta} values
for the Beta prior on \eqn{\gamma}, and optimal \eqn{\alpha_1} and
\eqn{\alpha_2} values for the Dirichlet prior on \eqn{\pi_{k,l}} when matching
state voter files over time.
}
\author{
Ben Fifield <benfifield@gmail.com>
}

