% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedupeMatches.R
\name{dedupeMatches}
\alias{dedupeMatches}
\title{dedupeMatches}
\usage{
dedupeMatches(matchesA, matchesB, EM,
matchesLink, patterns, linprog)
}
\arguments{
\item{matchesA}{A dataframe of the matched observations in
dataset A, with all variables used to inform the match.}

\item{matchesB}{A dataframe of the matched observations in
dataset B, with all variables used to inform the match.}

\item{EM}{The EM object from \code{emlinkMARmov()}}

\item{matchesLink}{The output from \code{matchesLink()}}

\item{patterns}{The output from \code{getPatterns()}.}

\item{linprog}{Whether to implement Winkler's linear programming solution to the deduplication
problem. Default is false.}
}
\value{
\code{dedupeMatches()} returns a list containing the following elements:
\item{matchesA}{A deduped version of matchesA}
\item{matchesB}{A deduped version of matchesB}
\item{EM}{A deduped version of the EM object}
}
\description{
Dedupe matched dataframes.
}
\author{
Ted Enamorado <ted.enamorado@gmail.com> and Ben Fifield <benfifield@gmail.com>
}
