% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emlinkMARmov.R
\name{emlinkMARmov}
\alias{emlinkMARmov}
\title{emlinkMARmov}
\usage{
emlinkMARmov(patterns, p.m = NULL, p.gamma.k.m = NULL, p.gamma.k.u = NULL,
  tol = NULL, iter.max = NULL, psi = NULL, mu = NULL, alpha0 = NULL,
  alpha1 = NULL, pos.ad = NULL)
}
\arguments{
\item{patterns}{table that holds the counts for each unique agreement
pattern. This object is produced by the function: tableCounts.}

\item{p.m}{probability of finding a match}

\item{p.gamma.k.m}{probability that conditional of being in the matched set we observed a specific agreement value for field k.}

\item{p.gamma.k.u}{probability that conditional of being in the non-matched set we observed a specific agreement value for field k.}

\item{tol}{convergence tolerance}

\item{iter.max}{Max Number of Iterations (5000 by default)}

\item{psi}{Value of psi parameter for beta prior on gamma}

\item{mu}{Value of mu parameter for beta prior on gamma}

\item{pos.ad}{Binary indicators for whether a given field is an address field. To be used when 'alpha0' and 'alpha1' are specified.
Default is 0 for all fields. Address fields should be set to 1 while non-address fields are set to 0.}
}
\description{
Expectation-Maximization algorithm for Record Linkage under the
Missing at Random (MAR) assumption.
}
\author{
Ted Enamorado <ted.enamorado@gmail.com> and Kosuke Imai
}

