% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastLink.R
\name{fastLink}
\alias{fastLink}
\title{fastLink}
\usage{
fastLink(dfA, dfB, varnames, stringdist.match,
partial.match = NULL,
cut.a, cut.p, n.cores = NULL, tol.em = 1e-04,
match = 0.85, verbose = FALSE)
}
\arguments{
\item{dfA}{Dataset A - to be matched to Dataset B}

\item{dfB}{Dataset B - to be matched to Dataset A}

\item{varnames}{A vector of variable names to use for matching.
Must be present in both dfA and dfB}

\item{stringdist.match}{A vector of booleans, indicating whether to use
string distance matching when determining matching patterns on
each variable. Must be same length as varnames.}

\item{partial.match}{A vector of booleans, indicating whether to include
a partial matching category for the string distances. Must be same length
as varnames. Default is NULL (FALSE for all variables).}

\item{cut.a}{Lower bound for full string-distance match, ranging between 0 and 1. Default is 0.92}

\item{cut.p}{Lower bound for partial string-distance match, ranging between 0 and 1. Default is 0.88}

\item{priors.obj}{A list containing priors for auxiliary movers information,
as output from calcMoversPriors(). Default is NULL}

\item{w.lambda}{How much weight to give the prior on lambda versus the data. Must range between 0 (no weight on prior) and 1 (weight fully on prior).
Default is NULL (no prior information provided).}

\item{w.pi}{How much weight to give the prior on pi versus the data. Must range between 0 (no weight on prior) and 1 (weight fully on prior).
Default is NULL (no prior information provided).}

\item{l.address}{The number of possible matching categories used for address fields. If a binary yes/no match, \code{l.address} = 2,
while if a partial match category is included, \code{l.address} = 3. Default is NULL (no prior information provided).}

\item{address.field}{A vector of booleans for whether a given field is an address field. To be used when 'pi.prior' is included in 'priors.obj'.
Default is FALSE for all fields. Address fields should be set to TRUE while non-address fields are set to FALSE.}

\item{gender.field}{A vector of booleans, indicating whether a matching variable
indicates gender. If so, the exact-matching gender prior is used in the EM algorithm.
Must be the same length as varnames if provided. Default is NULL (FALSE for all variables).}

\item{n.cores}{Number of cores to parallelize over. Default is NULL.}

\item{tol.em}{Convergence tolerance for the EM Algorithm. Default is 1e-04.}

\item{match}{A number between 0 and 1. The closer to 1 the more centainty you have about a given pair being a match}

\item{verbose}{Whether to print elapsed time for each step. Default is FALSE.}
}
\value{
\code{fastLink} returns a list of class 'fastLink' containing the following components:
\item{matches}{An nmatches X 2 matrix containing the indices of the successful matches in \code{dfA}
in the first column, and the indices of the corresponding successful matches in \code{dfB} in the
second column.}
\item{EM}{A matrix with the output of the EM algorithm, which contains the exact matching
patterns and the associated posterior probabilities of a match for each matching pattern.}
}
\description{
Run the fastLink algorithm to probabilistically match
two datasets.
}
\examples{
\dontrun{
fastLink(dfA, dfB, varnames = c("firstname", "lastname", "streetname", "birthyear"),
stringdist.match = c(TRUE, TRUE, TRUE, FALSE), partial.match = c(TRUE, TRUE, FALSE, FALSE),
verbose = TRUE)
}
}
\author{
Ted Enamorado <ted.enamorado@gmail.com>, Ben Fifield <benfifield@gmail.com>, and Kosuke Imai
}

