% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastLink.R
\name{fastLink}
\alias{fastLink}
\title{fastLink}
\usage{
fastLink(df_a, df_b, varnames, partial_match, tol.em = 1e-04,
  tol.match = 1e-07)
}
\arguments{
\item{df_a}{Dataset A - to be matched to Dataset B}

\item{df_b}{Dataset B - to be matched to Dataset A}

\item{varnames}{A vector of variable names to use for matching.
Must be present in both df_a and df_b}

\item{partial_match}{A vector of booleans, indicating whether to use
partial matching or exact matching when determining matching patterns on
each variable.}

\item{tol.em}{Convergence tolerance for the EM Algorithm. Default is 1e-04.}

\item{tol.match}{Convergence tolerance for determining matches. Default is 1e-07.}
}
\description{
Run the fastLink algorithm to probabilistically match
two datasets.
}

