% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammaCK2par.R
\name{gammaCK2par}
\alias{gammaCK2par}
\title{gammaCK2par}
\usage{
gammaCK2par(matAp, matBp, n.cores, cut.a, method, w)
}
\arguments{
\item{matAp}{vector storing the comparison field in data set 1}

\item{matBp}{vector storing the comparison field in data set 2}

\item{n.cores}{Number of cores to parallelize over. Default is NULL.}

\item{cut.a}{Lower bound for full match, ranging between 0 and 1. Default is 0.92}

\item{dedupe}{Logical. Whether internal linkage or linkage between two frames is being performed.}

\item{transform}{An optional function returning transformed strings or a list of transformed strings. 
First argument should be the strings to be transformed}

\item{transform.args}{An optional list of arguments to the transform function}

\item{method}{String distance method, options are: "jw" Jaro-Winkler (Default), "jaro" Jaro, and "lv" Edit. 
May also be passed as a custom function with argument list stringdist.args. The first two arguments of 
any custom function provided should be the strings to be compared}

\item{method.args}{List of arguments for custom function provided to method}

\item{w}{Parameter that describes the importance of the first characters of a string (only needed if method = "jw"). Default is .10}
}
\value{
\code{gammaCK2par} returns a list with the indices corresponding to each
matching pattern, which can be fed directly into \code{tableCounts} and \code{matchesLink}.
}
\description{
Field comparisons for string variables. Two possible agreement patterns are considered:
0 total disagreement, 2 agreement.
The distance between strings is calculated using a Jaro-Winkler distance.
}
\examples{
\dontrun{
g1 <- gammaCK2par(dfA$firstname, dfB$lastname)
}
}
\author{
Ted Enamorado <ted.enamorado@gmail.com>, Ben Fifield <benfifield@gmail.com>, and Kosuke Imai
}
