% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammaCKpar.R
\name{gammaCKpar}
\alias{gammaCKpar}
\title{gammaCKpar}
\usage{
gammaCKpar(matAp, matBp)
}
\arguments{
\item{matAp}{vector storing the comparison field in data set 1}

\item{matBp}{vector storing the comparison field in data set 2}
}
\description{
Field comparisons for string variables. Three possible agreement patterns are considered:
0 total disagreement, 1 somehow agree, 2 agreement.
The distance between strings is calculated using a Jaro-Winkler distance as implemented
in the stringdist package.
}
\author{
Ted Enamorado <ted.enamorado@gmail.com>, Ben Fifield, and Kosuke Imai
}

