% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammaKpar.R
\name{gammaKpar}
\alias{gammaKpar}
\title{gammaKpar}
\usage{
gammaKpar(matAp, matBp, gender, n.cores)
}
\arguments{
\item{matAp}{vector storing the comparison field in data set 1}

\item{matBp}{vector storing the comparison field in data set 2}

\item{gender}{Whether the matching variable is gender. Will override
standard warnings of missingness/nonvariability. Default is FALSE.}

\item{n.cores}{Number of cores to parallelize over. Default is NULL.}
}
\value{
\code{gammaKpar} returns a list with the indices corresponding to each
matching pattern, which can be fed directly into \code{tableCounts} and \code{matchesLink}.
}
\description{
Field comparisons: 0 disagreement, 2 total agreement.
}
\examples{
\dontrun{
g1 <- gammaKpar(dfA$birthyear, dfB$birthyear)
}
}
\author{
Ted Enamorado <ted.enamorado@gmail.com>, Ben Fifield <benfifield@gmail.com>, and Kosuke Imai
}

