% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPatterns.R
\name{getPatterns}
\alias{getPatterns}
\title{getPatterns}
\usage{
getPatterns(matchesA, matchesB, varnames, partial.match, gammalist)
}
\arguments{
\item{matchesA}{Indices of matches from dataframe A.}

\item{matchesB}{Indices of matches from dataframe B.}

\item{varnames}{A vector of variable names to use for matching.}

\item{partial.match}{A vector of booleans, indicating whether to include
a partial matching category for each column. Must be same length
as varnames. Default is FALSE for all variables.}

\item{gammalist}{A list of agreement pointers returned by gammaCKpar or similar functions for each column}
}
\value{
\code{getPatterns()} returns a dataframe with a row for each matched pair,
where each column indicates the matching pattern for each matching variable.
}
\description{
Get the full matching patterns for all matched pairs in dataset A and dataset B
}
\author{
Ted Enamorado <ted.enamorado@gmail.com> and Ben Fifield <benfifield@gmail.com>
}
