% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPosterior.R
\name{getPosterior}
\alias{getPosterior}
\title{getPosterior}
\usage{
getPosterior(matchesA, matchesB, EM, patterns)
}
\arguments{
\item{matchesA}{A dataframe of the matched observations in
dataset A, with all variables used to inform the match.}

\item{matchesB}{A dataframe of the matched observations in
dataset B, with all variables used to inform the match.}

\item{EM}{The EM object from \code{emlinkMARmov()}}

\item{patterns}{The output from \code{getPatterns()}.}
}
\value{
\code{getPosterior} returns the posterior probability of a match for each matched pair of observations
in matchesA and matchesB
}
\description{
Get the posterior probability of a match for each matched pair of observations
}
\author{
Ben Fifield <benfifield@gmail.com>
}
