% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchesLink.R
\name{matchesLink}
\alias{matchesLink}
\title{matchesLink}
\usage{
matchesLink(gammalist, nr1 = w, nr2 = p, em = y, cut = z,
  n.cores = NULL)
}
\arguments{
\item{gammalist}{A list of objects produced by either gammaKpar or
gammaCKpar.}

\item{nr1}{number of observations in dataset 1}

\item{nr2}{number of observations in dataset 2}

\item{em}{parameters obtained from the Expectation-Maximization algorithm under the MAR assumption. These estimates are
produced by emlinkMAR}

\item{cut}{is the interval of weight values for the agreements that we want to examine closer.}
}
\description{
matchesLink produces two dataframes that store
all the pairs that share a pattern that conforms
to the an interval of the Fellegi-Sunter
weights
}
\author{
Ted Enamorado <ted.enamorado@gmail.com>, Ben Fifield <benfifield@gmail.com>, and Kosuke Imai
}

