% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchesLink.R
\name{matchesLink}
\alias{matchesLink}
\title{matchesLink}
\usage{
matchesLink(gammalist, nr1, nr2, em, cut, n.cores = NULL)
}
\arguments{
\item{gammalist}{A list of objects produced by either gammaKpar or
gammaCKpar.}

\item{nr1}{number of observations in dataset 1}

\item{nr2}{number of observations in dataset 2}

\item{em}{parameters obtained from the Expectation-Maximization algorithm under the MAR assumption. These estimates are
produced by emlinkMAR}

\item{cut}{is the interval of weight values for the agreements that we want to examine closer.}

\item{n.cores}{Number of cores to parallelize over. Default is NULL.}
}
\value{
\code{matchesLink} returns an nmatches X 2 matrix with the indices of the
matches rows in dataset A and dataset B.
}
\description{
matchesLink produces two dataframes that store
all the pairs that share a pattern that conforms
to the an interval of the Fellegi-Sunter
weights
}
\examples{
\dontrun{
## Calculate gammas
g1 <- gammaCKpar(dfA$firstname, dfB$firstname)
g2 <- gammaCKpar(dfA$middlename, dfB$middlename)
g3 <- gammaCKpar(dfA$lastname, dfB$lastname)
g4 <- gammaKpar(dfA$birthyear, dfB$birthyear)

## Run tableCounts
tc <- tableCounts(list(g1, g2, g3, g4), nr1 = nrow(dfA), nr2 = nrow(dfB))

## Run EM
em <- emlinkMAR(tc)
EM <- data.frame(resultsEM$patterns.w)
EM$zeta.j <- resultsEM$zeta.j
EM <- EM[order(EM[, "weights"]), ]
EM$cumsum.m <- cumsum(EM[, "p.gamma.j.m"])
EM$cumsum.u <- 1 - cumsum(EM[, "p.gamma.j.u"])

match.ut <- EM$weights[ EM$zeta.j >= 0.85][1]

## Get matches
ml <- matchesLink(list(g1, g2, g3, g4), nr1 = nrow(dfA), nr2 = nrow(dfB),
em = em, cut = match.ut)
}

}
\author{
Ted Enamorado <ted.enamorado@gmail.com>, Ben Fifield <benfifield@gmail.com>, and Kosuke Imai
}
