% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_proc_gamma.R
\name{post_proc_gamma}
\alias{post_proc_gamma}
\title{post_proc_gamma}
\usage{
post_proc_gamma(dfA, dfB, varname, fastlinkres, gammalist, isoreg, method, method.args, 
transform, transform.args, n.cores, chunksize, resolution, min.posterior, drop.lowest)
}
\arguments{
\item{dfA}{First dataframe to be linked}

\item{dfB}{Second dataframe to be linked}

\item{varname}{Name of variable that posterior processing is based on. Defaults to 'Name'. Must be present in both dfA and dfB}

\item{fastlinkres}{Prior linkage result using exact matching on varname field, which gives fine-grained prior probabilities that any record pair is a match}

\item{gammalist}{Gammalist object containing sparser representation of field agreement from the prior linkage result}

\item{isoreg}{A compact format isotonic regression object, used to convert classifier scores to match probabilities under prior class imbalance}

\item{method}{String distance method, passed as a custom function with argument list stringdist.args. The first two arguments of 
any custom function provided should be the strings to be compared}

\item{method.args}{List of arguments for custom function provided to method}

\item{transform}{An optional function returning transformed strings or a list of string transformations. 
First argument should be the strings to be transformed}

\item{transform.args}{An optional list of arguments to the transform function}

\item{n.cores}{Number of cores to be used for parallel operations. Defaults to the number of available cores minus 1}

\item{chunksize}{Target number of record pairs that will be recovered and compared at a time. 
This parameter defaults to 90000, but may need to be adjusted for best performance, 
depending on the number of record pairs evaluated and the complexity of the fuzzy matching method}

\item{resolution}{1/Number of probability bins over which to aggregate results. Defaults to 1e-2 (100 bins)}

\item{unblocked.size}{Number of comparisons considered before resorting to parallel operations - can be tweaked to speed up performance in some scenarios}

\item{min.posterior}{Sets a cap to decide which agreement patterns are evaluated for fuzzy matching. Blocks for which the maximum attainable probability 
of match given the classifier score falls below this threshold are not evaluated. Defaults to 0.5.}

\item{drop.lowest}{Whether to drop matched indices from the lowest probability bin - this can save a lot of memory and processing wasted on nonmatches}

\item{keyvar}{optional variable name used to distinguish true match identity}

\item{cluster}{optionally provide a pre-made cluster object. can help with repeated cluster startup times.}

\item{diceroll.gamma}{if set to TRUE (the default), a gammalist-like object is constructed (binmatches) by sampling inclusion of each record pair 
according to the posterior probability of name matching, which provides an easy way to re-integrate the results into the fastLink protocol. If set to FALSE, 
binmatches is returned for ranges of posterior match scores binned according to the \code{resolution} parameter}
}
\value{
\code{post_proc_gamma} returns a list with the binned posterior counts, probabilities, and matching indices at each level but the lowest
}
\description{
posterior match probability calculation based on string similarity and Fellegi-Sunter priors
}
