% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMoversPriors.R
\name{precalcPriors}
\alias{precalcPriors}
\title{precalcPriors}
\usage{
precalcPriors(L, var.prior.gamma = NULL, var.prior.pi = NULL,
  gamma.mean = NULL, pi.mean = NULL)
}
\arguments{
\item{L}{Number of agreement categories for \eqn{\pi_{k,l}}. Default is NULL.}

\item{var.prior.gamma}{User-specified variance for the prior probability on gamma.}

\item{var.prior.pi}{User-specified variance for the prior probability on pi.}

\item{gamma.mean}{Prior mean for \eqn{\gamma}}

\item{pi.mean}{Prior mean for \eqn{pi_{k,l}}}
}
\author{
Ben Fifield <benfifield@gmail.com>
}

