% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableCounts.R
\name{tableCounts}
\alias{tableCounts}
\title{tableCounts}
\usage{
tableCounts(gammalist = NULL, nr1 = y, nr2 = z)
}
\arguments{
\item{gammalist}{A list of objects produced by either gammaKpar or
gammaCKpar.}

\item{nr1}{number of observations in dataset 1}

\item{nr2}{number of observations in dataset 2}
}
\description{
Count pairs with the same pattern in the cross product between two datasets.
}
\author{
Ted Enamorado <ted.enamorado@gmail.com> and Kosuke Imai
}

